:truck: moves mongo stuff into indexProgram
